feat(multisig): types for multisig sign
